requirements.txt
streamlit
requests
faiss-cpu
openai>=1.0.0
numpy
python-dotenv
